AnalysisData entries can be shared between jobs, so we can't delete them without checking if they're still used
At the moment, I can't think of a way of checking that without a
transaction / a locking mechanism, which could be slow in large databases
(cf hoover_pipeline.pl which doesn not hoover the analysis_data table for
the same reason).
Since I don't expect many input_ids to be changed this way, I think we can
accept keeping the old entries.